before refactor

+++ src/graph/FlowEdge.java
@@ -4,8 +4,8 @@ public class FlowEdge {
 	
 	private final int v;
 	private final int w;
-	private final int capacity;
-	private int flow;
+	private final double capacity;
+	private double flow;
 	
 	public FlowEdge(int v, int w, int capacity)
 	{
@@ -17,8 +17,8 @@ public class FlowEdge {
 	
 	public int from() {return v;}
 	public int to() {return w;}
-	public int capacity() {return capacity;}
-	public int flow() {return flow;}
+	public double capacity() {return capacity;}
+	public double flow() {return flow;}
 	
 	public int other(int vertex)
 	{
@@ -27,14 +27,14 @@ public class FlowEdge {
 		else throw new RuntimeException("Inconsistent Edge");
 	}
 	
-	public int residualCapacityTo(int vertex)
+	public double residualCapacityTo(int vertex)
 	{
 		if (vertex == v) return flow;
 		else if (vertex == w) return capacity - flow;
 		else throw new RuntimeException("Inconsistent Edge");
 	}
 	
-	public void addResidualFlowTo(int vertex, int delta)
+	public void addResidualFlowTo(int vertex, double delta)
 	{
 		if (vertex == v) flow -= delta;
 		else if (vertex == w) flow += delta;
